requirements.txt
biopython>=1.80
loguru>=0.7.0
pytest>=7.0.0
